Changelog and credits moved out of source directory.
This cleans up source a bit, but also allows github to see the files
better, as there system doesn't look in the src folder for key files.

Also, C99 style comments were removed the from the CREDITS and CHANGELOG
files, and the filenames got the *.md extension so that doxygen will
render them correctly.

paq.sh was updated to reflect this change, all seems nominal