Adds doxygen stubs to ALL FUNCTIONS
note: i have made no effort to actually document anything. I have simply
placed stub header blocks above all of the functions in the project. The
intent here would be to give new developers a template, along with
making it easier for anyone to fill in the documentation as the project
moves forward.