cleaned up header blocks in nuklear.h for doxygen output
i have taken the liberty to remove extra syntax in the header blocks
because doxygen does that for us. things like repeated functions names
and code blocks that used to format the old documentation system are
simply not necessary anymore. Doxygen will pull the function and
format it correctly without the need to copy that into the header block.