SFML demo not designed for automated build. Made it fail out early
The SFML OpenGL3 demo requires that the SFML and glad.h
files/directories be specified for the system that is running make. This
means that an automated system is unable to process the file, unless we
figure out some fancy way of automating the build of all GLAD, SFML, and
the demo. This is definitely possible; but too cumbersome for me to
figure out tonight. Instead, I just updated the Makefile in the demo to
bail out of the compilation if the SFML and GLAD directories have not
been specified. Hopefully this will prevent the github action from
failing out.